+++ src/service/pwsh.rs
@@ -1,9 +1,8 @@
-use std::env::temp_dir;
+use std::{env::temp_dir, process::Command};
 
 use itertools::Itertools;
-use tauri_plugin_shell::ShellExt;
 
-use crate::{error_handler::Result, seelen::get_app_handle, windows_api::WindowsApi};
+use crate::error::Result;
 
 const PWSH_COMMON_ARGS: [&str; 7] = [
     "-NoLogo",
@@ -18,7 +17,6 @@ const PWSH_COMMON_ARGS: [&str; 7] = [
 pub struct PwshScript {
     inner: String,
     args: Vec<String>,
-    elevated: bool,
 }
 
 impl PwshScript {
@@ -26,7 +24,6 @@ impl PwshScript {
         Self {
             inner: contents.into(),
             args: Vec::new(),
-            elevated: false,
         }
     }
 
@@ -39,41 +36,33 @@ impl PwshScript {
         self
     }
 
-    pub fn elevated(mut self) -> Self {
-        self.elevated = true;
-        self
-    }
-
     fn build_args(&self, script_path_str: &str) -> Vec<String> {
-        let mut args = PWSH_COMMON_ARGS
+        let args = PWSH_COMMON_ARGS
             .iter()
             .map(|s| s.to_string())
             .chain(["-File".to_string(), script_path_str.to_string()])
             .chain(self.args.clone())
             .collect_vec();
-        if self.elevated && !WindowsApi::is_elevated().unwrap_or(false) {
-            args = PWSH_COMMON_ARGS
-            .iter()
-            .map(|s| s.to_string())
-            .chain([
-                "-Command".to_string(),
-                format!("Start-Process 'powershell' -Verb runAs -WindowStyle Hidden -Wait -ArgumentList '{}'", args.join(" "))
-            ])
-            .collect_vec();
-        }
         args
     }
 
-    /// returns `Ok(stdout)` or `Err(stderr)`
-    ///
-    /// if elevated, will run as admin and always will return `Ok("")`
-    pub async fn execute(&self) -> Result<String> {
+    pub fn spawn(&self) -> Result<std::process::Child> {
+        let script_path = temp_dir().join(format!("slu-{}.ps1", uuid::Uuid::new_v4()));
+        std::fs::write(&script_path, &self.inner)?;
+        let args = self.build_args(&script_path.to_string_lossy());
+        let result = Command::new("powershell").args(args).spawn();
+        // std::fs::remove_file(&script_path)?;
+        Ok(result?)
+    }
+
+    /// returns `Ok(stdout)` or `Err(stderr || stdout)`
+    pub fn output(&self) -> Result<String> {
         let script_path = temp_dir().join(format!("slu-{}.ps1", uuid::Uuid::new_v4()));
         std::fs::write(&script_path, &self.inner)?;
 
         let args = self.build_args(&script_path.to_string_lossy());
-        let shell = get_app_handle().shell();
-        let result = shell.command("powershell").args(args).output().await;
+        let result = Command::new("powershell").args(args).output();
+
         // delete script before check output
         std::fs::remove_file(&script_path)?;
         let output = result?;